+++ Angular/index.ts
@@ -1,0 +1,5 @@
+export * from './roles-and-permissions.models';
+export * from './resource-matcher';
+export * from './roles-and-permissions.service';
+export * from './permission-visibility.directives';
+export * from './roles-and-permissions.module';